[Core] Extract a separate FallbackSummary class from GlutenImplicits
